generate modbus response to the master's request

+++ plc_1xx_L433/Src/Flash_manager.c
@@ -93,7 +93,7 @@ uint16_t crc16(uint8_t *adr_buffer, uint32_t byte_cnt)
 }
 
 
-uint8_t read_registers_from_flash(uint32_t* data_out)
+uint8_t read_registers_from_flash(uint16_t* data_out)
 {	
 	uint32_t flash_set[REG_COUNT+1];
 	uint32_t orig_crc = 0;
@@ -120,7 +120,7 @@ uint8_t read_registers_from_flash(uint32_t* data_out)
 }
 
 
-uint8_t write_registers_to_flash(uint32_t* data)
+uint8_t write_registers_to_flash(uint16_t* data)
 {	
 	uint32_t flash_set[REG_COUNT+1];
 	uint32_t crc = 0;
