+++ plc_1xx_L433/Src/Flash_manager.h
@@ -12,13 +12,13 @@ typedef struct
     uint32_t crc;	
 } flash_data;
 
-uint8_t read_registers_from_flash(uint32_t* data_out);
-uint8_t write_registers_to_flash(uint32_t* data);
+uint8_t read_registers_from_flash(uint16_t* data_out);
+uint8_t write_registers_to_flash(uint16_t* data);
 uint16_t crc16(uint8_t *adr_buffer, uint32_t byte_cnt);
 
 //Массив для хранения настроек
-uint32_t settings[REG_COUNT];
-uint32_t default_settings[REG_COUNT] = {10,10,10,10,};
+uint16_t settings[REG_COUNT];
+uint16_t default_settings[REG_COUNT] = {10,10,10,10,};
